% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrvgo.R
\name{calculateSimMatrix}
\alias{calculateSimMatrix}
\title{calculateSimMatrix
Calculate the score similarity matrix between terms}
\usage{
calculateSimMatrix(
  x,
  orgdb,
  semdata = GOSemSim::godata(orgdb, ont = ont),
  ont = c("BP", "MF", "CC"),
  method = c("Resnik", "Lin", "Rel", "Jiang", "Wang")
)
}
\arguments{
\item{orgdb}{one of org.* Bioconductor packages (the package name, or the
package itself)}

\item{semdata}{object with prepared GO DATA for measuring semantic similarity}

\item{ont}{ontlogy. One of c("BP", "MF", "CC")}

\item{method}{distance method. One of the supported methods by GOSemSim:
c("Resnik", "Lin", "Rel", "Jiang", "Wang")}

\item{goterms}{vector of GO terms}
}
\value{
a square matrix with similarity scores between terms
}
\description{
calculateSimMatrix
Calculate the score similarity matrix between terms
}
\examples{
go_analysis <- read.delim(system.file("extdata/example.txt", package="rrvgo"))
simMatrix <- calculateSimMatrix(go_analysis$ID, orgdb="org.Hs.eg.db", ont="BP", method="Rel")
}
