% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrvgo.R
\name{getGoSize}
\alias{getGoSize}
\title{getGoSize
Get GO term size (# of genes)}
\usage{
getGoSize(terms, orgdb)
}
\arguments{
\item{terms}{GO terms}

\item{orgdb}{one of org.* Bioconductor packages (the package name, or the
package itself)}
}
\value{
number of genes associated with each term
}
\description{
getGoSize
Get GO term size (# of genes)
}
