% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrvgo.R
\name{getGoTerm}
\alias{getGoTerm}
\title{getGoTerm
Get the description of a GO term}
\usage{
getGoTerm(x)
}
\arguments{
\item{x}{GO terms}
}
\value{
the Term slot in GO.db::GOTERM[[x]]
}
\description{
getGoTerm
Get the description of a GO term
}
