% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrvgo.R
\name{getTermDisp}
\alias{getTermDisp}
\title{getTermDisp
Calculate the term dispensability score, defined as the semantic similarity
threshold a term was assigned to a cluster (namely, the similarity of a term
to the cluster representative term).}
\usage{
getTermDisp(simMatrix, cluster, clusterRep)
}
\arguments{
\item{simMatrix}{a (square) similarity matrix}

\item{cluster}{the cluster assignment for each term}

\item{clusterRep}{the cluster representative term}
}
\value{
a vector of term dispensability scores
}
\description{
getTermDisp
Calculate the term dispensability score, defined as the semantic similarity
threshold a term was assigned to a cluster (namely, the similarity of a term
to the cluster representative term).
}
