% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrvgo.R
\name{getTermUniq}
\alias{getTermUniq}
\title{getTermUniq
Calculate the term uniqueness score, defined as 1 minus the average semantic
similarity of a term to all other terms.}
\usage{
getTermUniq(simMatrix, cluster = NULL)
}
\arguments{
\item{simMatrix}{a (square) similarity matrix}

\item{cluster}{vector with the cluster each entry in the simMatrix belongs to.
If NULL, a}
}
\value{
a vector of term uniqueness scores
}
\description{
getTermUniq
Calculate the term uniqueness score, defined as 1 minus the average semantic
similarity of a term to all other terms.
}
