% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlib.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue
Emulate ggplot2 color palette.}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
a vector with colors (alphanumeric)
}
\description{
gg_color_hue
Emulate ggplot2 color palette.
}
\details{
It is just equally spaced hues around the color wheel, starting from 15:
}
\examples{
\dontrun{
plot(1:10, pch=16, cex=2, col=gg_color_hue(10))
}
}
