% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVGO.R
\name{loadOrgdb}
\alias{loadOrgdb}
\title{loadOrgdb
Load an orgdb object}
\usage{
loadOrgdb(orgdb)
}
\arguments{
\item{orgdb}{one of org.* Bioconductor packages}
}
\value{
the loaded orgdb
}
\description{
loadOrgdb
Load an orgdb object
}
