% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prepareDemoData}
\alias{prepareDemoData}
\title{prepareDemoData}
\usage{
prepareDemoData()
}
\value{
Whatever the call to write.table() would return
}
\description{
Prepare demo data included in the package.
}
\details{
Taken asis from https://yulab-smu.github.io/clusterProfiler-book/chapter3.html#input-data,
which at the same time it is taken from the R package breastCancerMAINZ. It
contains 200 samples with breast cancer at different grades (I, II and III).
The dataset basically contains log2 ratios of the geometric means of
grade III vs. grade I samples.
}
