% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVGO.R
\name{reduceSimMatrix}
\alias{reduceSimMatrix}
\title{reduceSimMatrix
Reduce a set of GO terms based on their semantic similarity and scores.}
\usage{
reduceSimMatrix(simMatrix, scores = NULL, threshold, orgdb)
}
\arguments{
\item{simMatrix}{a (square) similarity matrix}

\item{scores}{*named* vector with scores (weights) assigned to each term.
Higher is better. Can be NULL (default, means no scores). Note: if you have
p-values as scores, consider -1*log-transforming them (`-log(p)`)}

\item{threshold}{similarity threshold}

\item{orgdb}{one of org.* Bioconductor packages (the package name, or the
package itself)}
}
\value{
a data.frame with all terms and it's "reducer" (NA if the term was not reduced)
}
\description{
reduceSimMatrix
Reduce a set of GO terms based on their semantic similarity and scores.
}
\details{
Remove terms with a similarity higher than `threshold`. Decide which term
remains based on a score. If no score is provided, the select either the
broader or the narrower one (`untie` parm).
}
\examples{
go_analysis <- read.delim(system.file("extdata/example2.txt", package="RVGO"))
scores <- setNames(-log10(go_analysis$V2), go_analysis$V1)
simMatrix <- calculateSimMatrix(go_analysis$V1, orgdb="org.Hs.eg.db", ont="BP", method="Rel")
reduced_go_analysis <- reduceSimMatrix(simMatrix, scores, threshold=0.7, orgdb="org.Hs.eg.db")
}
