% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_rrvgo.R
\name{shiny_rrvgo}
\alias{shiny_rrvgo}
\title{shiny_rrvgo
Launch an interactive web interface.}
\usage{
shiny_rrvgo(...)
}
\arguments{
\item{...}{other params sent to shiny::runApp().}
}
\value{
Nothing
}
\description{
shiny_rrvgo
Launch an interactive web interface.
}
