% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_RVGO.R
\name{shiny_RVGO}
\alias{shiny_RVGO}
\title{shiny_RVGO}
\usage{
shiny_RVGO(...)
}
\arguments{
\item{...}{other params sent to shiny::runApp()}
}
\description{
Launch an interactive web interface
}
