% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrvgo.R
\name{getGoSize}
\alias{getGoSize}
\title{getGoSize
Get GO term size (# of genes)}
\usage{
getGoSize(terms, orgdb, keytype, children)
}
\arguments{
\item{terms}{GO terms}

\item{orgdb}{one of org.* Bioconductor packages (the package name, or the
package itself)}

\item{keytype}{keytype passed to AnnotationDbi::keys to retrieve GO terms 
associated to gene ids in your orgdb}

\item{children}{include genes in children terms (based on relationships in
the GO DAG hierarchy)}
}
\value{
number of genes associated with each term
}
\description{
getGoSize
Get GO term size (# of genes)
}
